Implement parsing of SonarQube rule overrides
Create new model class that is being read by the SonarQubeRuleDefinitionOverrideXmlFileParser, which is an implementation of the SAX parser just like the other parsers already implemented. The parser will read from an XML input stream and create a collection of model instances that can be accessed through the implementation of the SonarQubeRuleDefinitionOverrideProvider interface.