Create base language class and derivatives for C# and VB.NET
Create a new abstract base class for languages that does most of the work for registering new languages within SonarQube. Update the classes CSharpLanguage and VBNetLanguage to make use of this new base class, cleaning up their implementations a bit and removing duplicated code.

IMHO both languages are not required to be added to SonarQube explicitly, because there are already official SonarQube plugins that will introduce these languages. This plugin should simply rely on their presence or somehow require them in order to work.