Improve XML data validation
Create new simple interface to validate XML data. Create abstract base class for XML data validation which requires an implementation of this interface in order to validate the XML data correctly. Most of the code has been extracted from the previous version of the InspectCodeXmlFilValidator class, which is now merely a wrapper to generate a Validator instance for InspectCode report files using the corresponding XML Schema Definition file. The same is true for the new class SonarQubeRuleDefinitionOverrideXmlFileValidator, which is used to validate the integrity of the SonarQube rule override definition files.