Register extensions for VisualBasic.NET (rule repo and sensor)
Register the new VBNetRulesDefinition and VBNetSensor classes with the supplied SonarQube context in order to make them available within SonarQube.

IMHO both language classes (CSharpLanguage and VBNetLanguage) should not be added to SonarQube context explicitly, because there are already official SonarQube plugins that will introduce these languages. This plugin should simply rely on their presence or somehow require them in order to work. Not sure how the latter would work, though as there is no useful documentation for this.