Update Sensor description and describe where it is used
Change description of Sensor classes to a more concise definition of where they are originated. Typically, the description is printed into the console output of the SonarScanner for MSBuild in a list of plugins that will be downloaded from SonarQube and executed after the build job. It's more like a descriptive name than a description of the purpose of the plugin.