Import rules for severity "DO_NOT_SHOW".
These are just disabled by default and depending on the project/company policy they should be enabled or not. Previously these types of rules were not imported at all, making them unavailable for tracking in SonarQube.
This was noticeable due to the fact that we now export the ReSharper rules via InspectCode with --no-buildin-settings to make them even with the ReSharper default. Some rules (e.g. TabsAreDisallowed) changed in that process to "DO_NOT_SHOW" which made them disappear after an update to this plugin >= 1.0.0.202013.

With this change we now support/provide C# rules with "DO_NOT_SHOW" severity because this is just a pre-set that the individual use could change and otherwise these rules would not even be available.

Issue #6